+++ src/main/kotlin/com/ibrahimdans/i18n/plugin/ide/folding/FoldingBuilderBase.kt
@@ -79,6 +79,7 @@ abstract class FoldingBuilderBase(private val lang: Lang) : FoldingBuilderEx(), 
                         }
                 }
             }
+            .sortedBy { it.range.startOffset }
             .toTypedArray()
     }