+++ src/test/kotlin/com/ibrahimdans/i18n/extensions/localization/yaml/CodeHighlightingTest.kt
@@ -4,8 +4,9 @@ import com.ibrahimdans.i18n.plugin.ide.runWithConfig
 import com.ibrahimdans.i18n.plugin.ide.settings.Config
 import com.ibrahimdans.i18n.plugin.utils.generator.code.JsCodeGenerator
 import com.ibrahimdans.i18n.plugin.utils.generator.translation.YamlTranslationGenerator
-import com.intellij.testFramework.fixtures.BasePlatformTestCase
+import com.ibrahimdans.i18n.plugin.PlatformBaseTest
 import com.intellij.testFramework.fixtures.CodeInsightTestFixture
+import org.junit.jupiter.api.Test
 
 private fun CodeInsightTestFixture.customCheck(fileName: String, code: String, translationName: String, translation: String) = this.runWithConfig(
     Config(defaultNs = "translation")
@@ -15,10 +16,11 @@ private fun CodeInsightTestFixture.customCheck(fileName: String, code: String, t
     this.checkHighlighting(true, true, true, true)
 }
 
-class CodeHighlightingTest2: BasePlatformTestCase() {
+class CodeHighlightingTest2: PlatformBaseTest() {
     val cg = JsCodeGenerator()
     val tg = YamlTranslationGenerator()
 
+    @Test
     fun testReferenceToObject() = myFixture.customCheck(
             "refToObject.${cg.ext()}",
             cg.generate("\"test:<error descr=\"Reference to object\">tst2.plurals</error>\""),
@@ -26,6 +28,7 @@ class CodeHighlightingTest2: BasePlatformTestCase() {
             tg.generatePlural("tst2", "plurals", "value", "value1", "value2", "value5")
     )
 
+    @Test
     fun testExpressionInsideTranslation() = myFixture.customCheck(
             "expressionInTranslation.${cg.ext()}",
             cg.generate("isSelected ? \"test:<error descr=\"Reference to object\">tst2.plurals</error>\" : \"test:<error descr=\"Unresolved key\">unresolved.whole.key</error>\""),
@@ -33,6 +36,7 @@ class CodeHighlightingTest2: BasePlatformTestCase() {
             tg.generatePlural("tst2", "plurals", "value", "value1", "value2", "value5")
     )
 
+    @Test
     fun testResolved() = myFixture.customCheck(
             "resolved.${cg.ext()}",
             cg.generate("\"test:tst1.base.single\""),
@@ -40,6 +44,7 @@ class CodeHighlightingTest2: BasePlatformTestCase() {
             tg.generatePlural("tst2", "plurals", "value", "value1", "value2", "value5")
     )
 
+    @Test
     fun testReferenceToObjectDefaultNs() = myFixture.customCheck(
             "refToObjectDefNs.${cg.ext()}",
             cg.generate("\"<error descr=\"Reference to object\">tst2.plurals</error>\""),
@@ -47,6 +52,7 @@ class CodeHighlightingTest2: BasePlatformTestCase() {
             tg.generatePlural("tst2", "plurals", "value", "value1", "value2", "value5")
     )
 
+    @Test
     fun testNotArg() = myFixture.customCheck(
             "defNsUnresolved.${cg.ext()}",
             cg.generateInvalid(