+++ src/test/kotlin/com/ibrahimdans/i18n/plugin/ide/actions/ExtractionCancellationTest.kt
@@ -9,6 +9,7 @@ import com.intellij.openapi.ui.TestDialogManager.setTestDialog
 import com.intellij.openapi.ui.TestDialogManager.setTestInputDialog
 import com.intellij.openapi.ui.TestInputDialog
 import org.junit.jupiter.api.Assertions.*
+import org.junit.jupiter.api.Test
 
 class ExtractionCancellationTest: PlatformBaseTest() {
 
@@ -22,6 +23,7 @@ class ExtractionCancellationTest: PlatformBaseTest() {
 
     private val testJson = JsonTranslationGenerator().generateContent("ref", "section", "key", "Reference in Json")
 
+    @Test
     fun testCancel() {
         myFixture.configureByText("simple.js", simpleJs)
         myFixture.addFileToProject("assets/test.json", testJson)
@@ -36,6 +38,7 @@ class ExtractionCancellationTest: PlatformBaseTest() {
         myFixture.checkResult("assets/test.json", testJson, false)
     }
 
+    @Test
     fun testCancelInvalid() {
         myFixture.configureByText("simple.js", simpleJs)
         myFixture.addFileToProject("assets/test.json", testJson)
@@ -56,6 +59,7 @@ class ExtractionCancellationTest: PlatformBaseTest() {
         myFixture.checkResult("assets/test.json", testJson, false)
     }
 
+    @Test
     fun testExtractionUnavailable() {
         myFixture.configureByText(
             "unavailable.tsx",
@@ -73,6 +77,7 @@ class ExtractionCancellationTest: PlatformBaseTest() {
         assertEquals(emptyList<IntentionAction>(), myFixture.filterAvailableIntentions(hint).toList())
     }
 
+    @Test
     fun testExtractionUnavailable2() {
         myFixture.configureByText(
             "unavailable.tsx",